requirements.txt
Flask 
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
Flask-Cors
Flask-api
mysql-connector-python
Python-Decouple
Requests
Autopep8
simple-salesforce

